Various tweaks -  description
Tweaks on the send page and wallet page.

- Removed coin logo on mobile version and left label present

- Added a small header with the coin name under the label to give a hint
of what coin the user is looking at.

- Receive and send fixed, next to each other on desktop and one under
the other on mobile. Added the highlight on the "Send" while on mobile.

- Added another input for the currency ( FIAT and Coin ).
Had to move the small label indicating the currency to the left, the
input type="number" arrows are already placed on the right.